Move Climate*Button components to Common namespace
Refactored ClimateButton, ClimateDropdownButton, and ClimateLinkButton components and their code-behind files from the .Common.Control namespace to .Common. Updated all usages and imports to the new namespace across the codebase. Restored original markup and parameters for these components. Made minor formatting and whitespace adjustments.
