+++ ClimateExplorer.Web.Client/Components/Chart/ChartAxisListView.razor
@@ -1,4 +1,5 @@
 @using Blazorise
+@using ClimateExplorer.Web.Client.Components.Common
 
 <div class="axes-dropdown-wrapper">
     <ClimateDropdownButton Text="Axes" Icon="fas fa-chart-line">
