+++ ClimateExplorer.Web.Client/Components/Common/ClimateButton.razor.cs
@@ -1,4 +1,4 @@
-namespace ClimateExplorer.Web.Client.Components.Common.Control;
+namespace ClimateExplorer.Web.Client.Components.Common;
 
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
