+++ ClimateExplorer.Web.Client/Components/DataSetBrowser.razor
@@ -1,10 +1,4 @@
-@using ClimateExplorer.Core.Model;
-@using ClimateExplorer.Core.ViewModel
-@using ClimateExplorer.Web.UiModel
-@using Blazorise
-@using Blazorise.Components
-@using Blazorise.Charts
-@using static ClimateExplorer.Core.Enums
+@using ClimateExplorer.Web.Client.Components.Common
 
 @if (RootFolders != null)
 {
