+++ ClimateExplorer.Web.Client/Components/MapContainer.razor
@@ -1,7 +1,4 @@
-@using Blazorise
-@using ClimateExplorer.Core.Model;
-@using CurrentDevice;
-@using DPBlazorMapLibrary;
+@using ClimateExplorer.Web.Client.Components.Common
 
 <div class="mapContainer collapsed" id="mapContainer">
     @if (mapOptions != null)
