+++ ClimateExplorer.Web.Client/Pages/HomeOverviewInfo.razor
@@ -1,4 +1,5 @@
 @namespace ClimateExplorer.Web.Client.Components
+@using ClimateExplorer.Web.Client.Components.Common
 
 <p>
     ClimateExplorer helps you explore how the climate is changing in your region and around the world. There are two main sections: this page and the regional &amp; global section. Try clicking elements such as <ClimateLinkButton Text="record high" Class="action-link" OnClick="OnShowRecordHigh" />.
