+++ ClimateExplorer.Web.Client/_Imports.razor
@@ -8,6 +8,6 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using ClimateExplorer.Web.Client
-@using ClimateExplorer.Web.Client.Components.Common.Control
+@using ClimateExplorer.Web.Client.Components.Common
 @using ClimateExplorer.Core.Blog
 @using CurrentDevice
